Add files via upload
Communities and Crime:
Using LASSO, Boosting for Regression, PCR model to predict the probability of crime 
APS Failure data:
Using Tree-Based Methods to determined APS. The data set consists of data collected from heavy Scania trucks in everyday usage. The system in focus is the Air Pressure system (APS) which generates air pollution that are utilized in various functions in a truck, such as braking and gear changes.